login and register page completed
