+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
+    "fetch-intercept": "^2.3.1",
     "react": "^16.10.2",
     "react-alert": "^5.5.0",
     "react-dom": "^16.10.2",
