+++ src/CSS/signupComponent.css
@@ -1,0 +1,36 @@
+
+.lbl {
+  margin-top: 25px;
+  float: right
+}
+.input {
+  margin-top: 20px;
+  float: left;
+}
+.clear {
+  margin-top: 20px;
+  float: left;
+}
+.heading{
+  margin-left: 10%;
+  color: red;
+  font-weight: bold;
+  margin-bottom: 40px;
+}
+.message{
+  color: green;
+  margin-top: 50px;
+}
+
+.homeimage{
+  width:100%;
+  height: 10%;
+  padding-left: 10%;
+  padding-right: 10%;
+  padding-bottom: 10%;
+}
+.home{
+  color:blue;
+  text-align: center;
+  margin-top: 50px;
+}
