--- src/CSS/userComponent.css
@@ -1,44 +1,0 @@
-.user{
-
-    margin-left: -300px;
-    margin-top: 50%;
-}
-.gender{
-    margin-left: -20%;
-}
-.users{
-    margin-left: 10%;
-    
-}
-.form{
-    margin-top: 100%;
-}
-.gform{
-    text-align: left;
-}
-.form{
-    margin-top: 2%;
-}
-.lbl {
-    margin-top: 25px;
-    float: right
-}
-.input {
-    margin-top: 20px;
-    float: left;
-}
-.clear {
-    margin-top: 20px;
-    float: left;
-}
-.heading{
-    margin-left: 10%;
-    color: red;
-    font-weight: bold;
-    margin-bottom: 40px;
-}
-
-.message{
-    color: green;
-    margin-top: 50px;
-}
