+++ src/Reducer/alertreducer.js
@@ -9,6 +9,7 @@ import {
     error: '',
   }
   const alert = (state = initialState, action) => {
+    debugger
     switch (action.type) {
       case SUCCESS_ALERT:
         return {
