+++ src/Reducer/helperreducer.js
@@ -1,7 +1,4 @@
-import{
-    API_REQUEST_PENDING,
-    API_REQUEST_COMPLETE
-  } from '../Action/helper.action'
+import{ API_REQUEST_PENDING, API_REQUEST_COMPLETE } from '../Action/helper.action'
   
   const initialState = {
     pending: false
