+++ src/Reducer/index.js
@@ -1,18 +1,15 @@
 import {combineReducers}from 'redux';
 import { reducer as formReducer } from 'redux-form'
-import userReducer from './userReducer';
-import helper from './helper.reducer';
-import alert from './alert.reducer';
+import signupReducer from './signupReducer';
+import helper from './helperreducer';
+import alert from './alertreducer';
 import signinReducer from './signinReducer';
 
-
 export default combineReducers({
-     
-     form: formReducer,
-     userReducer,
-     helper,
-     alert,
-     signinReducer
-   
+    form: formReducer,
+    signupReducer,
+    helper,
+    alert,
+    signinReducer   
 });
