+++ src/Reducer/signinReducer.js
@@ -1,36 +1,25 @@
-import {
-    SIGNIN_REQUEST_SUCCESS,
-    SIGNIN_REQUEST_FAILURE
-  } from '../Action/signinAction';
-  
+import { SIGNIN_REQUEST_SUCCESS, SIGNIN_REQUEST_FAILURE } from '../Action/signinAction';
     
   const initialState = {
     email: '',
-    password:'',
-    message:''
+    password:''
   }
-  
-    
+     
   const signIn = (state = initialState, action) => {
-      debugger
     switch(action.type){
       case SIGNIN_REQUEST_SUCCESS:
       return{
         ...state,
-        pending: false,
-        message:action.resp
+        pending: false
       }
       case SIGNIN_REQUEST_FAILURE:
       return{
         ...state,
-        pending: false,
-        // errormessage:action.error
-    
-
+        pending: false
+           
       }
       default:
       return state
     }
   }
-  
   export default signIn
