+++ src/index.js
@@ -4,8 +4,7 @@ import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import Reducer from './Reducer';
-import ReduxThunk from 'redux-thunk'
-	
+import ReduxThunk from 'redux-thunk'	
 import { Provider } from 'react-redux';	
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/css/bootstrap.css';
